Fixed out-of-phase sampling issue
When the plugin auto-detects CPOL/CPHA
switch (mainly for simulation), the logic
does not ensure that it checks the polarity at
the point where the enable line goes low.
This results in certain cases where the CPHA
is 180 degrees out-of-phase, leading to
potential (very likely in real world) sampling errors.