Resolved plugin crashing
In certain cases, the enable-line's current sample
position could be tailing behind the current position
being processed on the clock line. When using this
enable channel as the input argument to
WouldAdvancingToAbsPositionCauseTransition() this
would result in invalid behavior per the Analyzer
SDK and will cause the software to crash as the
analyzer is technically being asked to check if it can
advanced forward from a position that is behind its
current position. To avoid this invalid case, the
logic has been updated to check the sample position and
auto-fast-forward the enable-line's sample position to
that of the clock line to avoid this issue.