Rewrite of simulation behavior
Simulation now demonstrates AFSI communication.
A file containing the plugins version is continuously read out.
This provides a simple way for metadata, that only exists in
the Windows DLL, to be available for other operating systems.

Simulation also provides basic excitation of the process data
by feeding it with sinusoidal waveform data points.