Fixed advanced settings XML behavior
The state of the variables that are controlled via the
advanced settings XML file were not being interfaced
with the text archive. This means that the settings would
be lost after the software closes and reopens a capture.

With this change, the captures' archive will now contain
the state of the XML file settings instead of using the
default values.