Added support for simulating dynamic msg lengths
Specifying a value < 0 for "SpiMessageDataLength" in
the Advanced Settings XML file will active this functionality.

This also addresses a "last fragment" issue in log file sim. Message
header bytes were not being included in the calculation of
mTotalMsgDataBytesToSend. This could cause abnormal behavior
in certain cases. To help clarify the point of
"mTotalMsgDataBytesToSend" this variable has been renamed to
"mTotalMsgBytesToSend" since this specifies the entire length of the
message not just the message data.