Fixed network object exception logic
The exception handler logic was not checking if the
network object's "exception information" attribute was
being accessed before returning true when the settings
network type was "unspecified". This would cause the
logic to generate output as if there was an exception
and in turn report to the user the attribute value could
not be enumerated since no network type was specified.